+++ KeyboardVisualizerCommon/LEDStrip.cpp
@@ -125,7 +125,7 @@ void LEDStrip::SetLEDs(COLORREF pixels[64][256])
 
         for (int idx = 0; idx < (num_leds * 3); idx += 3)
         {
-            int col_idx = (idx * (255.0f / (3.0f * num_leds))) + (128.0f / (3.0f * num_leds));
+            int col_idx = (int)((idx * (255.0f / (3.0f * num_leds))) + (128.0f / (3.0f * num_leds)));
             serial_buf[idx + 1] = GetRValue(pixels[ROW_IDX_BAR_GRAPH][col_idx]);
             serial_buf[idx + 2] = GetGValue(pixels[ROW_IDX_BAR_GRAPH][col_idx]);
             serial_buf[idx + 3] = GetBValue(pixels[ROW_IDX_BAR_GRAPH][col_idx]);
